requirements.txt
babel==2.16.0
numpy==2.1.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tkcalendar==1.6.1
tzdata==2024.1
tzlocal==5.2